+++ apps/fabric-website/src/pages/Controls/PivotPage/PivotPage.tsx
@@ -4,8 +4,7 @@ import { ControlsAreaPage, IControlsPageProps } from '../ControlsAreaPage';
 import { PivotPageProps } from './PivotPage.doc';
 import { Platforms } from '../../../interfaces/Platforms';
 
-const baseUrl =
-  'https://github.com/OfficeDev/office-ui-fabric-react/tree/master/apps/fabric-website/src/pages/Controls/PivotPage/';
+const baseUrl = 'https://github.com/microsoft/fluentui/tree/master/apps/fabric-website/src/pages/Controls/PivotPage/';
 
 export const PivotPage: React.FunctionComponent<IControlsPageProps> = props => {
   return (