+++ packages/charting/src/components/DonutChart/examples/DonutChart.Basic.Example.tsx
@@ -20,6 +20,13 @@ export class DonutChartBasicExample extends React.Component<IDonutChartProps, {}
       chartTitle: chartTitle,
       chartData: points,
     };
-    return <DonutChart data={data} innerRadius={30} href={'https://developer.microsoft.com/en-us/'} />;
+    return (
+      <DonutChart
+        data={data}
+        innerRadius={30}
+        href={'https://developer.microsoft.com/en-us/'}
+        legendsOverflowText={'overflow Items'}
+      />
+    );
   }
 }