+++ packages/fluentui/react-bindings/src/index.ts
@@ -15,11 +15,16 @@ export { default as unstable_useDispatchEffect } from './hooks/useDispatchEffect
 export { default as useIsomorphicLayoutEffect } from './hooks/useIsomorphicLayoutEffect';
 export { default as useStateManager } from './hooks/useStateManager';
 export { default as useStyles } from './hooks/useStyles';
+export { default as useUnhandledProps } from './hooks/useUnhandledProps';
 
 export { default as unstable_createAnimationStyles } from './styles/createAnimationStyles';
 export { default as unstable_calculateAnimationTimeout } from './styles/calculateAnimationTimeout';
 export { default as unstable_getStyles } from './styles/getStyles';
 export * from './styles/types';
+
+export { default as createSvgIcon } from './icons/createSvgIcon';
+export * from './icons/createSvgIcon';
+export * from './icons/types';
 
 export { default as useTelemetry } from './telemetry/useTelemetry';
 export * from './telemetry/types';