+++ packages/fluentui/react-northstar/src/themes/teams/componentStyles.ts
@@ -46,6 +46,7 @@ export { default as HierarchicalTreeItem } from './components/HierarchicalTree/h
 export { default as HierarchicalTreeTitle } from './components/HierarchicalTree/hierarchicalTreeTitleStyles';
 
 export { default as Icon } from './components/Icon/iconStyles';
+export { default as SvgIcon } from './components/Icon/svgIconStyles';
 
 export { default as Image } from './components/Image/imageStyles';
 
@@ -115,3 +116,10 @@ export { default as CarouselNavigationItem } from './components/Carousel/carouse
 export { default as Table } from './components/Table/tableStyles';
 export { default as TableRow } from './components/Table/tableRowStyles';
 export { default as TableCell } from './components/Table/tableCellStyles';
+
+export { default as Card } from './components/Card/cardStyles';
+export { default as CardHeader } from './components/Card/cardHeaderStyles';
+export { default as CardBody } from './components/Card/cardBodyStyles';
+export { default as CardFooter } from './components/Card/cardFooterStyles';
+export { default as CardPreview } from './components/Card/cardPreviewStyles';
+export { default as CardTopControls } from './components/Card/cardTopControlsStyles';