+++ packages/fluentui/react-northstar/test/specs/components/Card/Card-test.tsx
@@ -1,0 +1,10 @@
+import { isConformant, handlesAccessibility } from 'test/specs/commonTests';
+import Card from 'src/components/Card/Card';
+
+describe('Card', () => {
+  isConformant(Card, { constructorName: 'Card' });
+
+  describe('accessiblity', () => {
+    handlesAccessibility(Card, { defaultRootRole: 'group' });
+  });
+});