+++ packages/example-data/CHANGELOG.md
@@ -1,7 +1,13 @@
 # Change Log - @uifabric/example-data
 
-This log was last generated on Fri, 17 Jan 2020 02:32:17 GMT and should not be manually modified.
+This log was last generated on Mon, 23 Mar 2020 12:28:29 GMT and should not be manually modified.
 
+## 7.0.8
+Mon, 23 Mar 2020 12:28:29 GMT
+
+### Patches
+
+- Replace OfficeDev/office-ui-fabric-react with microsoft/fluentui (elcraig@microsoft.com)
 ## 7.0.3
 Fri, 17 Jan 2020 02:32:17 GMT